[Fly][A - Quest] LK 68-80 [Kick] & [Fly][H - Quest] 68-80 [Kick]
Quest The Last Line of Defense had a problem where toons would frequently not make it into the cannon. I found that the cannon shooting was too erratic. Changed this cannon to mimic the cannon use in the DK starter zone profile by adding gravity and velocity behavior attributes to the profile. Also fixed an issue with the bot trying to turn in the quest when it wasn't finished.